requirements.txt
asgiref==3.11.0
Django==5.1.4
gunicorn==23.0.0
packaging==25.0
python-dateutil>=2.8.2
sqlparse==0.5.4
tzdata==2025.3
whitenoise==6.11.0
pytesseract>=0.3.10
Pillow>=10.0.0
